requirements.txt
mysql-connector-python
openai
azure-search-documents
python-dotenv